server/requirements.txt
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
langchain==0.1.13
python-dotenv==1.0.1
openai==1.14.3
langchain-community==0.0.29
langchain-openai==0.1.1
unstructured==0.12.6
Markdown==3.6
pypdf==4.1.0
pdf2image==1.17.0
gunicorn==21.2.0
